-JDBC实现对数据的增删改查(下)

+++ java-web-foundational/src/main/java/net/microsoft/java/web/foundational/dao/impl/StatementUserDaoImplV2.java
@@ -1,7 +1,6 @@
 package net.microsoft.java.web.foundational.dao.impl;
 
-import com.mysql.cj.jdbc.Driver;
-import net.microsoft.java.web.foundational.dao.Dao;
+import net.microsoft.java.web.foundational.dao.UserDao;
 import net.microsoft.java.web.foundational.entity.User;
 import net.microsoft.java.web.foundational.util.JDBCUtil;
 
@@ -15,7 +14,7 @@ import java.util.List;
  * @author:Suche
  **/
 
-public class StatementDaoImplV2 implements Dao {
+public class StatementUserDaoImplV2 implements UserDao {
 
     static String quotes = "'";
     static String comma = ",";
@@ -31,8 +30,12 @@ public class StatementDaoImplV2 implements Dao {
         String sql = "insert into jdbc_user values(null," + quotes + user.getName() + quotes + comma + quotes + user.getPassword() + quotes + comma + "now(),now()" + ")";
         final Statement statement;
         try {
-            statement = connection.createStatement();
-            statement.executeUpdate(sql);
+
+            if (sql != null && sql != "") {
+                statement = connection.createStatement();
+                final int row = statement.executeUpdate(sql);
+                return row;
+            }
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
@@ -43,15 +46,18 @@ public class StatementDaoImplV2 implements Dao {
     public int delete(User user) {
         String sql = null;
         try (
-                Connection connection = DriverManager.getConnection(url, userName, userPassword);
+                Connection connection = JDBCUtil.getConnection();
                 Statement statement = connection.createStatement()
+
         ) {
-            if (user.getId() != null && user != null) {
-                sql = "delete from jdbc_user where id = " + quotes + user.getId() + quotes;
-                System.out.println("sql = " + sql);
-                if (null != sql && sql != "") {
-                    int row = statement.executeUpdate(sql);
-                    return row;
+            if (sql != null && sql != "") {
+                if (user.getId() != null && user != null) {
+                    sql = "delete from jdbc_user where id = " + quotes + user.getId() + quotes;
+                    System.out.println("sql = " + sql);
+                    if (null != sql && sql != "") {
+                        int row = statement.executeUpdate(sql);
+                        return row;
+                    }
                 }
             } else if (null != user && user.getName() != null) {
                 sql = "delete from jdbc_user where name = " + quotes + user.getName() + quotes;
@@ -72,7 +78,7 @@ public class StatementDaoImplV2 implements Dao {
         String sql = null;
         List<User> userList = new ArrayList<>();
         try (
-                final Connection connection = DriverManager.getConnection(url, userName, userPassword);
+                final Connection connection = JDBCUtil.getConnection();
                 final Statement statement = connection.createStatement()
         ) {
             if (user == null) {
@@ -118,5 +124,10 @@ public class StatementDaoImplV2 implements Dao {
         }
 
         return null;
+    }
+
+    @Override
+    public int update(User user) {
+        return 0;
     }
 }
